Update paper.md
Affiliation update and some suggested edits, which are mostly stylistic. 
Could maybe mention that with a snake config for your HPC you can run jobs in parallel and it's much faster, but IDK if anyone cares haha. 
Otherwise, reads really clearly and it's great to better understand what you've done!